Add files via upload
Sensors used: 
   To find out the amount of groceries left: Ultrasonic sensor, Load cell
   To gas leakage and fire accidents if any: MQ-6
Database used: Firebase
Technologies: Internet of Things

Working:  The real time readings of sensor(analog values) are calculated from arduino code and this data is continuously sent to a serial port of the system. That serial data is read by the python client and it(python application) takes care of pushing the data to cloud. This data from cloud can be retrieved into a mobile application.

Applications: Day-to-day analysis of usage, alerts is there is shortage of sugar or something, generates a shopping list automatically... :)